tools/py/pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
